Move login to dedicated charger/vaillant package
